Added script to do all operations one step at a time
With this script it will instead do all the operations from shuffled fastq file to sorted BAM file one at a time. It is configured to use a temporary folder somewhere else - in this case, a tmpfs RAM drive to severely speed up I/O. It should be possible to put this in a job scheduler, same as any other script. Just create multiple lists of files (or even one file at a time per list) and send it to SLURM or condor.

Note that before usage you'll need to edit the paths and, potentially, the tools it uses! It is all held in simple variables.